# SANGAM-PRODUCTION/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SANGAM — Supply Chain Management API",
    "version": "1.0.0",
    "description": "REST API for the SANGAM permissioned blockchain-based supply chain\nmanagement system built for the Indian Army.\n\n## Authentication\nAll protected endpoints require a `Bearer` JWT obtained from `POST /api/auth/login`.\nTokens expire after 8 hours (one operational shift). Use `POST /api/auth/refresh` to rotate.\n\n## Permissions\nEvery endpoint lists the required permission (format: `resource:action`).\nPermissions are granted by role (Day 13 RBAC), temporary delegation (Day 15),\nor single-use emergency override (Day 15).\n\n## Command Scope\nData is filtered to the requesting user's command scope — their unit and all\nsubordinate units. A battalion CO automatically sees all companies and platoons.\nLateral access (sibling units) is denied.\n\n## Offline-First\nAll services degrade gracefully when the database is unavailable.\nNotifications and audit events are buffered in memory and flushed on reconnect.",
    "contact": {
      "name": "SANGAM Development Team",
      "url": "https://github.com/sangam-scm"
    },
    "license": {
      "name": "Proprietary — Indian Army Internal Use"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local / Docker development"
    },
    {
      "url": "https://sangam.army.mil",
      "description": "Production (classified network)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /api/auth/login (Day 14). Expires after 8h."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "INVALID_REQUEST"
          },
          "message": {
            "type": "string",
            "example": "A descriptive error message"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions or outside command scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Invalid request body or parameters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/reports/dashboard": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Command dashboard (all reports)",
        "description": "Runs all 6 report types for the current user's command scope in parallel. Cached 5 minutes per user. Add ?refresh=true to bypass cache. Requires reports:read.",
        "operationId": "getDashboard",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "refresh",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard payload — stock, transfers, blockchain, mesh, security, roster",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/reports/stock-levels": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Stock level report",
        "description": "Per-unit/category inventory totals and low-stock items across the user's command scope. Requires reports:read.",
        "operationId": "getStockLevels",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stock report",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/reports/security-posture": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Security posture report",
        "description": "Combines Day 13 audit event counts (SECURITY/CRITICAL) with Day 11 pending acknowledgments. Requires reports:advanced (SENIOR_OFFICER+).",
        "operationId": "getSecurityPosture",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "hours",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 24
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Security posture",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/reports/export/{type}": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Export report as CSV",
        "description": "Download a report as CSV. Supported types: stock-levels, transfers, unit-roster, mesh-health. Requires reports:export.",
        "operationId": "exportReport",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "stock-levels",
                "transfers",
                "unit-roster",
                "mesh-health"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "security": []
}